initial db refactor
