+++ src/screens/db-views/Databases.jsx
@@ -6,8 +6,8 @@ import Fab from '@mui/material/Fab'
 import withReactContent from 'sweetalert2-react-content'
 import { useStateValue, actions } from '../state'
 import { getAllDatabases, removeDatabase, createDatabase } from '../database'
-import { ProgramList } from '../db-components/DatabaseList'
-import CreateDialog from '../db-components/CreateDialog'
+import { ProgramList } from '../components/db-components/DatabaseList'
+import CreateDialog from '../components/db-components/CreateDialog'
 import { ProjectDetailsContext } from '../ProjectDetailsContext'
 import Backdrop from '../components/templates/Backdrop'
